--- advent_of_code/aoc/2022/inputs/day14_sample.txt
@@ -1,2 +1,0 @@
-498,4 -> 498,6 -> 496,6
-503,4 -> 502,4 -> 502,9 -> 494,9
